Преобразованиетипа переменной s2

+++ src/main/java/laba4/CalcPDF.java
@@ -245,11 +245,10 @@ public class CalcPDF extends HttpServlet implements Inter1,Inter2 {
 				double s1=i1.xsum3(sum5, sum4, z, h); //int s1=(int)((((sum5)+(sum4)-((sum4)*z)))*h);
 			
 				
-				int s2=(int)((s1*p*g*r));//-(s1*r)); //	int s2=(int)(s1*p*g-s1*r);
+				double s2=((s1*p*g*r));//-(s1*r)); //	int s2=(int)(s1*p*g-s1*r); 	int s2=(int)((s1*p*g*r));
 				 // gen=Integer.toString(s2);
 				  request.setAttribute("result5", s2);
 		
-			
 		
 			request.setAttribute("Srok", TextSrok);
 			request.setAttribute("Usl1", TextUsl1);
@@ -265,7 +264,6 @@ public class CalcPDF extends HttpServlet implements Inter1,Inter2 {
 			request.setAttribute("Usl11", TextUsl11);
 			request.setAttribute("Usl12", TextUsl12);
 			request.setAttribute("Promo", TextPromo);
-			
 			request.setAttribute("Kfpromo", TextKfpromo);
 			request.setAttribute("col", Textcol);
 			request.setAttribute("col1", Textcol1);
